requirements.txt
requests
python-dotenv
streamlit
pymupdf
huggingface-hub
scikit-learn